// src/types/celebrity.ts
// export interface Celebrity {
//   id: number;
//   first: string;
//   last: string;
//   dob: string;
//   gender: string;
//   email: string;
//   picture: string;
//   country: string;
//   description: string;
// }

// export type GenderOption = 'Male' | 'Female' | 'Transgender' | 'Rather not say' | 'Other';


  
export interface Celebrity {
  id: number;
  first: string;
  last: string;
  dob: string;
  gender: string;
  email: string;
  picture: string;
  country: string;
  description: string;
}

export type GenderOption = 'Male' | 'Female' | 'Transgender' | 'Rather not say' | 'Other';